requirements.txt
streamlit==1.17.0
python-dotenv==0.21.0
groq==0.1.0 